HIVSeqinR_ver2.6
This pipeline processes HIV near-full-genome sequences and categorize genomes into "intact" or into various "defective" categories.  For each sequence input, this pipeline will label it with a single verdict.  There are 31 possible verdicts highlighted in yellow below (including “Intact,” “PrematureStop,” “Hypermut” etc).  When a verdict contains the words “Check,” it is a flag to alert users that there is an abnormality.  In these cases, I suggest users to manually validate that the pipeline’s verdict is correct.  The final MyVerdict for each sequence input can be found in the last column in output file "Output_MyBigSummary_DF_FINAL.csv."